(MSVC) Fix warnings

+++ deps/rzlib/ioapi.c
@@ -190,7 +190,7 @@ static long ZCALLBACK fseek64_file_func (voidpf  opaque, voidpf stream, ZPOS64_T
     }
     ret = 0;
 
-    if(fseek((FILE *)stream, offset, fseek_origin) != 0)
+    if(fseek((FILE *)stream, (long)offset, fseek_origin) != 0)
                         ret = -1;
 
     return ret;
